refactor(document): give `internal::Document` the read-only defaults nine engines pasted
`is_editable` → false, `is_savable` → false and the two throwing `save`
overloads were written out verbatim in csv, iwork, markdown, rtf, the three
`oldms` documents and the ooxml spreadsheet and presentation documents. Only
odf and ooxml text have anything of their own to say, so the base class carries
the read-only answer and the nine copies go.

No behaviour change. Step 1 of #770.
