+++ src/odr/internal/markdown/markdown_document.cpp
@@ -34,21 +34,6 @@ const StyleRegistry &Document::style_registry() const {
   return m_style_registry;
 }
 
-bool Document::is_editable() const noexcept { return false; }
-
-bool Document::is_savable(const bool encrypted) const noexcept {
-  (void)encrypted;
-  return false;
-}
-
-void Document::save(std::ostream & /*out*/) const {
-  throw UnsupportedOperation();
-}
-
-void Document::save(std::ostream & /*out*/, const char * /*password*/) const {
-  throw UnsupportedOperation();
-}
-
 namespace {
 
 class ElementAdapter final : public abstract::ElementAdapter,
