+++ src/odr/internal/oldms/presentation/ppt_document.hpp
@@ -18,12 +18,6 @@ public:
 
   [[nodiscard]] const StyleRegistry &style_registry() const;
 
-  [[nodiscard]] bool is_editable() const noexcept override;
-  [[nodiscard]] bool is_savable(bool encrypted) const noexcept override;
-
-  void save(std::ostream &out) const override;
-  void save(std::ostream &out, const char *password) const override;
-
 private:
   ElementRegistry m_element_registry;
   StyleRegistry m_style_registry;
