+++ src/odr/internal/rtf/rtf_document.cpp
@@ -33,21 +33,6 @@ const ElementRegistry &Document::element_registry() const {
   return m_element_registry;
 }
 
-bool Document::is_editable() const noexcept { return false; }
-
-bool Document::is_savable(const bool encrypted) const noexcept {
-  (void)encrypted;
-  return false;
-}
-
-void Document::save(std::ostream & /*out*/) const {
-  throw UnsupportedOperation();
-}
-
-void Document::save(std::ostream & /*out*/, const char * /*password*/) const {
-  throw UnsupportedOperation();
-}
-
 namespace {
 
 class ElementAdapter final : public abstract::ElementAdapter,
